// src/generated/models/Cr720_wfrd3dobjs_2sModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr720_wfrd3dobjs_2sstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr720_wfrd3dobjs_2sstatecode = keyof typeof Cr720_wfrd3dobjs_2sstatecode;
export const Cr720_wfrd3dobjs_2sstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr720_wfrd3dobjs_2sstatuscode = keyof typeof Cr720_wfrd3dobjs_2sstatuscode;

export interface Cr720_wfrd3dobjs_2sBase {
  cr720_equipment_description?: string;
  cr720_id?: string;
  cr720_pl?: string;
  cr720_wfrd3dobjs_2id: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr720_wfrd3dobjs_2sstatecode;
  statuscode?: Cr720_wfrd3dobjs_2sstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr720_wfrd3dobjs_2s extends Cr720_wfrd3dobjs_2sBase {
  cr720_attachments?: string;
  cr720_attachments_name?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Cr720_wfrd3dobjs_2sFileColumnName = 'cr720_attachments';

export type Cr720_wfrd3dobjs_2sUploadColumnName = Cr720_wfrd3dobjs_2sFileColumnName;